delete mapbox and change to leafleat
