+++ client/src/pages/MapPage.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Header from '../components/Header.jsx';
-import Map from '../components/Map.jsx';
+// import Map from '../components/Map.jsx';
 
 function MapPage({ gasInfoHead, title, fuelValue }) {
   return (
     <>
       <Header title={title} gasInfoHead={gasInfoHead} fuelValue={fuelValue} />
-      <Map />
+      {/* <Map /> */}
     </>
   );
 }
