+++ package.json
@@ -18,7 +18,6 @@
   "license": "ISC",
   "dependencies": {
     "concurrently": "^7.1.0",
-    "mapbox-gl": "^2.8.2",
     "react-icons": "^4.3.1"
   }
 }
